// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/workallocationsummary.ts
export class WorkAllocationSummary {
    userId: string;
    userName: string;
    workAllocated: string;
    minWork: string;
    maxWork: string;
    maxDeadLineDate: string;
    isSupport: string;
    color: string;
    pageNumber: number;
    pageSize: string;
    skip: number;
    sortBy: String;
    sortDirectionAsc: boolean;
    SearchText: string;
    teamLeadId:string;
    entityId: string;
    isReportingOnly:boolean;
    isAll: boolean;
    isMyself: boolean;
    projectId: string;
}